+++ spec/jwt/jwk/decode_with_jwk_spec.rb
@@ -162,23 +162,6 @@ RSpec.describe JWT do
           )
         end
       end
-
-      if defined?(RbNaCl)
-        context 'when OKP keys are used' do
-          before do
-            skip('Requires the rbnacl gem') unless JWT.rbnacl?
-          end
-
-          let(:keypair) { RbNaCl::Signatures::Ed25519::SigningKey.new(SecureRandom.hex) }
-          let(:algorithm) { 'ED25519' }
-
-          it 'decodes the token' do
-            key_loader = ->(_options) { JSON.parse(JSON.generate(public_jwks)) }
-            payload, _header = described_class.decode(signed_token, nil, true, { algorithms: [algorithm], jwks: key_loader })
-            expect(payload).to eq(token_payload)
-          end
-        end
-      end
     end
   end
 end